Fixes #796 This fixes a bug introduced during the warning cleanups where a variable was set to be 16bit instead of 32bit
git-svn-id: http://svn.mythtv.org/svn/trunk@8291 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/programs/mythtranscode/replex/ringbuffer.h
@@ -40,7 +40,7 @@ extern "C" {
 	typedef struct ringbuffer {
 		int read_pos;
 		int write_pos;
-		uint16_t size;
+		uint32_t size;
 		uint8_t *buffer;
 	} ringbuffer;
